fix(deps): fix incompatibility with ktor's telemetry plugin
In newer versions of some part of opentelemetry, there's a
classcastexception caused internally by the javaagent and local otel
version differing Enough:tm:. This commit bumps those versions so
da-opentelemetry-javaagent works with opentelemetry 1.52.

Also bumped apache-commons-lang3 while I was at it, since there's an
open security advisory on it, and it was the easier fix of the two
libraries.