fix(config): default service name to env variable if present
With the previous setup, the serviceName supplied to the config service
and the name supplied to jaeger or similar differed when exclusively
using environment variables.

This commit also changes the readOnly default to false, as it doesn't
really make sense for the default to be true; the reason for the
configuration service is to use dynamic config, so it should really be
opt-out rather than opt-in, especially considering reverting readOnly
requires restarting the entire configuration service.